Avoid leaving behind stale L3Outs when clearing gateway
For pre-existing L3Outs, if the gateway for router
connecting to corresponding Neutron external network
is removed, we were re-creating the L3Out as part
of unsetting contract. This was because the unset
call wasn't be scoped to the correct tenant.

Signed-off-by: Amit Bose <bose@noironetworks.com>
(cherry picked from commit a807e92516819d6b803d204ec277d86bb34b47f6)