Fixing port binding for SNAT ports
By, (1) populating the binding information in the DB
and not using the driver's bind_port method

(2) populate the device_id of the port with the host name
and use it as a filter criteria. This also fixes the issue
that the SNAT port was only getting created on the first host
and not subsequently (since the 'binding:host' filter criteria
was being silently ignored).