Fixing port binding for SNAT ports
By, (1) populating the binding information in the DB
and not using the driver's bind_port method

(2) populate the device_id of the port with the host name
and use it as a filter criteria. This also fixes the issue
that the SNAT port was only getting created on the first host
and not subsequently (since the 'binding:host' filter criteria
was being silently ignored).

(cherry picked from commit 5f3df35a884fce8f7c33853a3be6fdb0ebf994c2)