Prevent unnecessary error message for SNAT network
Whenever a get_network operation is performed in the admin
context, the ML2 plugin returns the details of the DB-only SNAT
network(s). While constructing the network details the following
method is invoked:
https://github.com/openstack/neutron/blob/6c38103e095dbf107bdb199979be8e5e909d7e74/neutron/plugins/ml2/managers.py#L123-L130

which logs an error if there is no segment associated with a network.

The SNAT network does not require a segment, however we create a
local segment in this patch to comply with the above ML2 requirement
and avoid the error message in the logs.

(cherry picked from commit 1c81197213bc421d844b30df0e89134e27abdd9a)