Support allocating host SNAT IPs from user created subnets
The host SNAT auto IP allocation feature creates a SNAT subnet
on a SNAT network based on the CIDR in the host_pool_cidr configuration.
However, if this CIDR is exhausted, there is no way to add more subnet(s).

With this patch, the user will be able to create additional subnet(s) on
the SNAT network (the ones which have a name prefix
"host-snat-network-for-internal-use-"). Once this subnet is created
successfully, the host SNAT IPs will continue to be allocated from the newly
added subnet(s).

Note that these user defined subnet(s) should always have a gateway IP set,
so the "no-gateway" option should not be chosen when creating them.