Send correct host SNAT IP in get_dbp_details
We look at the binding-host of a port to decide the
host SNAT IP that should be returned in GBP details.
For a VM that is in the process of live migration,
the binding-host may still be the source host while
the RPC request comes from the destination host. In
this case we wrongly send the host SNAT IP of the
source host.
The fix is to use the host of the agent making the
RPC call to decide which host SNAT IP to report.

Signed-off-by: Amit Bose <bose@noironetworks.com>