Remove L2 plugin parent class.
This removes the NeutronDbPluginV2 class from the
inheritance hierarchy, and changes from the DVR
parent class to the non-DVR calss. It replaces
references to the ML2 core plugin with the member
variable defined in the superclass.  It also addresses
some problems with the l3 plugin unit tests (still
can't run them individually using tox, but they will
run when .tox/py27/bin/activate is sourced and
python -m testtools.run is used).

Signed-off-by: Thomas Bachman <tbachman@yahoo.com>