Added DVS VIF type support.
This modifies the mechanism driver to support port
binding for DVS VIF types. Since mechanism drivers
qualify port binding by only binding for hosts that
are running a matching agent, it adds a constraint
that an OpFlex agent and DVS agent can't exist on
the same host (otherwise, binding would be ambiguous).

Signed-off-by: Thomas Bachman <tbachman@yahoo.com>