Added DVS VIF type support.
This modifies the mechanism driver to support port
binding for DVS VIF types. Since mechanism drivers
qualify port binding by only binding for hosts that
are running a matching agent, it adds a constraint
that an OpFlex agent and DVS agent can only exist
on the same host when the OpFlex agent is used for
network services, such as dhcp and metadata agents,
and not compute services (otherwise, binding of ports
for compute would be ambiguous).

Signed-off-by: Thomas Bachman <tbachman@yahoo.com>