Handle multiple routers with NAT properly
If a tenant uses multiple routers, removing
any one of them was disrupting connectivity
through the other routers. This is because
we are removing the shadow L3-out when any
router is removed/gateway is unset. Instead
now we remove the shadow when the tenant is
not connected to the external network through
any router.

Signed-off-by: Amit Bose <amitbose@gmail.com>
(cherry picked from commit fa32ebb5fddcecb4f94e63913cb0f5577c630c65)