Allow connecting non-virtual nodes to OpFlex networks
Currently, OpFlex virtual networks support only virtual
endpoints created on KVM hosts. With this change,
non-virtual nodes, like service appliances, can also be
configured to be part of the same network that VMs
are connected to. This is achieved by statically binding
specific switch ports in ACI fabric to the corresponding EPG
using VLAN-encapsulation. For this we use hierarchical
port-binding - when Neutron ports bind to particular hosts,
the mechanism driver creates a dynamic VLAN
segment and allows other drivers to complete the binding
(if no other mechanism drivers complete the binding, we
complete the binding for the dynamic segment).
Static path-binding is configured on APIC when port-binding
is successful in Neutron.

Closes noironetworks/support#113

Signed-off-by: Amit Bose <amitbose@gmail.com>