[VRF per router] Tie VRF lifetime to that of the router
In vrf-per-router mode, when a router is created/deleted,
the corresponding VRF will also be created/removed.
This allows us to keep the code simple instead
of creating the VRF when the router is used and
removing it when the router is not being used.

Closes noironetworks/support#253

Signed-off-by: Amit Bose <amitbose@gmail.com>
(cherry picked from commit 4d51dfdf10e92cb1a4be1d46c8013568cc6de9f2)