Add support for L2 only mode
This adds support for using the ACI fabric only
for L2. The ACI-specific L3 plugin is not used,
as L3 is handled external to the fabric (e.g. using
the reference implementation/namespace router). The
code auto-detects if an ACI-specific L3 plugin is
configured, except in the case where the Cisco Router
Service L3 Plugin with ACI integration is used. If
the Cisco Router Service L3 Plugin, and L3 is to be
handled by the fabric, then the new l3_cisco_router_plugin flag
must be set to true, in order to ensure that L2 only mode isn't
enabled/used (note: it's possible to auto-detect
this as well, but that behavior is not implemented
in this patch).

Closes noironetworks/support#350

Signed-off-by: Thomas Bachman <tbachman@yahoo.com>
(cherry picked from commit 8e000f054ecd3ed178214b7cad0114e9342d24cc)