NameAlias support for tenant/app_profile
1. will set the nameAlias field when the 1st network is being
created under the tenant. This is also the time when the tenant
is really getting created in apic.
2. once the tenant is already in apic, we will update the nameAlias
field whenver the corresponding project name has been changed on
OS side. We do this through keystone notifications.

TODO: still need to put the keystone exchange and topic in our config
file as those are configurable in keystone.conf. Will do this in a
separate patch then.