Use new session for name mapping calls
Caching the session in the object doesn't
work with Neutron because of session pooling
- the cached session may be use in another
thread while ApicDbModel is using it. This
results in SQLAlchemy errors.
I'm just fixing the name-mapping methods
for now because that's where problems
were detected in tests. Other methods should
also be fixed to not use cached session.

Signed-off-by: Amit Bose <amitbose@gmail.com>