Fix keystone v3
The keystone v3 integration was broken. This adds an additional
section to the configuration file, apic_aim_auth, which allows
the specification of credential information needed by apicapi,
so that it can look up the openstack projects in keystone.

The old style of configuration is also kept, in order to preserve
backwards-compatibility.