Fix session use for AAPs
The port_ha_ipaddress_binding module is used by the legacy
plugin as well as the unified plugin's stable/newton branch.
Support for use of CIDRs in AAPs was added to the unified
plugin, but not the legacy plugin, which required an API
change for this module. This patch fixes the missing API
change.